requirements.txt
fastapi
uvicorn
pillow
jinja2
